add usersId check for all services and controllers

+++ src/services/contacts.js
@@ -29,8 +29,9 @@ export const getAllContacts = async ({
 }) => {
   const skip = perPage * (page - 1);
 
-  const contactsQuery = Contact.find();
-  contactsQuery.where('userId').equals(userId);
+  const contactsQuery = Contact.find({ userId });
+  // contactsQuery.where('userId').equals(userId);
+
   if (isFavourite) {
     contactsQuery.where('isFavourite').equals(isFavourite);
   }
@@ -53,8 +54,8 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactById = async (id) => {
-  return await Contact.findById(id);
+export const getContactById = async (id, userId) => {
+  return await Contact.findById({ _id: id, userId });
 };
 
 export const createContact = async (payload, userId) => {
@@ -62,8 +63,8 @@ export const createContact = async (payload, userId) => {
   return contact;
 };
 
-export const deleteContact = async (id) => {
-  return await Contact.findByIdAndDelete(id);
+export const deleteContact = async (id, userId) => {
+  return await Contact.findByIdAndDelete({ _id: id, userId });
 };
 
 export const updateContact = async (
